part1(old files) of everythng done: js design is on.

+++ carts/models.py
@@ -6,7 +6,7 @@ from products.models import GlobalPrice, SaleList
 class GlobalCart(models.Model):
 	sales_list = models.ForeignKey(SaleList, on_delete = models.CASCADE, related_name = 'slist_by')
 	gp_product = models.ForeignKey(GlobalPrice, on_delete = models.CASCADE, related_name = 'gp_by')
-	qnt = models.IntegerField()
-
+	qnt = models.PositiveSmallIntegerField()
+	
 	def __str__(self):
 		return f'{self.sales_list} : {self.gp_product.product.title} at {self.gp_product.price} * {self.qnt}'
